Adds support for durable upperdirs and keeping runtimes around
Adds --keep-runtime argument to spfs run, and spk commands. The
keep_runtime value is stored in the runtime's data and used by spfs
enter and monitor.

Using --keep-runtime will make the runtime use a durable upperdir root
path based on the runtime's name, and will stop it being deleted when
the process exits. The upper root paths are stored in the local repo
in durables directory to keep them together and because they must not
be on NFS for overlayfs.

Durable runtimes can only removed by spfs runtime rm and prune when
'--remove-durable' is given, and deletion includes the removal of
the durable upper root path for a durable runtime.

+++ crates/spfs-cli/cmd-enter/src/cmd_enter.rs
@@ -162,11 +162,13 @@ impl CmdEnter {
 
             let mut monitor_stdin = match spfs::monitor::spawn_monitor_for_runtime(&owned) {
                 Err(err) => {
-                    if let Err(err) = owned.delete().await {
-                        tracing::error!(
-                            ?err,
-                            "failed to cleanup runtime data after failure to start monitor"
-                        );
+                    if !owned.keep_runtime() {
+                        if let Err(err) = owned.delete().await {
+                            tracing::error!(
+                                ?err,
+                                "failed to cleanup runtime data after failure to start monitor"
+                            );
+                        }
                     }
                     return Err(err.into());
                 }
